refactor(preprocessing): extract text normalization and book metadata preprocessing to separate module
- Extracted normalize_text, subjects_to_string, preprocess_book, and create_embedding_input functions into a new module 'preprocessing.py'.
- Updated generate_embeddings.py to import preprocessing functions from preprocessing.py.
- Replaced hardcoded absolute paths with relative paths using pathlib for improved portability.
- Enhanced logging and error handling throughout the pipeline.
- Separated concerns into distinct functions for better readability and maintainability.

This refactor modularizes the embedding generation process and improves overall code quality.
